+++ docs/CHANGELOG.md
@@ -20,6 +20,8 @@ We generally fold multiple commits pertaining to the same topic as a single entr
 - Read the `Breaking Changes` section (here in the changelog).
 - Please report any issue!
 
+## VERSION 0.8.x
+
 ## VERSION 0.8.21
 
 ### Breaking Changes